Create To-do_ListV0.0.py
This is one of my first "beginner Python" posts on Github. I figured it a good idea to get early ideas here for a number of reasons. I want to build more interesting and robust versions of my ideas as I learn more Python, but I also want to have my work available to other new programmers, so they might be inspired to do the same ^^