To-do_ListV0.0
This is my first Python submission, in the form of a basic to-do list. I hope to evolve this into an actual user interface as I learn more Python. I figured it a good idea that I start posting all my little beginner projects, as they could be useful for others just starting out :)